+++ Common/src/main/java/fuzs/puzzleslib/common/api/client/data/v2/AbstractModelProvider.java
@@ -63,10 +63,35 @@ public abstract class AbstractModelProvider implements DataProvider {
     /**
      * @see #generateForBlocks(BlockModelGenerators, BlockSetFamily, Map)
      */
+    @Deprecated(forRemoval = true)
     public static Map<BlockSetVariant, BiConsumer<BlockModelGenerators, Block>> createVariantWoodBlockProviders(BlockSetFamily blockSetFamily, Block strippedBlock) {
+        return createVariantWoodBlockProviders(blockSetFamily);
+    }
+
+    /**
+     * @see #generateForBlocks(BlockModelGenerators, BlockSetFamily, Map)
+     */
+    public static Map<BlockSetVariant, BiConsumer<BlockModelGenerators, Block>> createVariantWoodBlockProviders(BlockSetFamily blockSetFamily) {
         return ImmutableMap.<BlockSetVariant, BiConsumer<BlockModelGenerators, Block>>builder()
+                .put(BlockSetVariant.LOG, (BlockModelGenerators blockModelGenerators, Block block) -> {
+                    blockModelGenerators.woodProvider(block).logWithHorizontal(block);
+                })
+                .put(BlockSetVariant.WOOD, (BlockModelGenerators blockModelGenerators, Block block) -> {
+                    blockModelGenerators.woodProvider(blockSetFamily.getBlockVariants()
+                            .get(BlockSetVariant.LOG)
+                            .value()).wood(block);
+                })
+                .put(BlockSetVariant.STRIPPED_LOG, (BlockModelGenerators blockModelGenerators, Block block) -> {
+                    blockModelGenerators.woodProvider(block).logWithHorizontal(block);
+                })
+                .put(BlockSetVariant.STRIPPED_WOOD, (BlockModelGenerators blockModelGenerators, Block block) -> {
+                    blockModelGenerators.woodProvider(blockSetFamily.getBlockVariants()
+                            .get(BlockSetVariant.STRIPPED_LOG)
+                            .value()).wood(block);
+                })
                 .put(BlockSetVariant.SHELF, (BlockModelGenerators blockModelGenerators, Block block) -> {
-                    blockModelGenerators.createShelf(block, strippedBlock);
+                    blockModelGenerators.createShelf(block,
+                            blockSetFamily.getBlockVariants().get(BlockSetVariant.STRIPPED_LOG).value());
                 })
                 .build();
     }